RF: move fips-related data from disclosure_backend to fips app.
