+++ disclosure_backend/project/serializers.py
@@ -5,7 +5,3 @@ from rest_framework import serializers
 class ContributionSerializer(serializers.ModelSerializer):
     class Meta:
         model = RcptCd
-
-
-class LocationSerializer(serializers.Serializer):
-    pass
