+++ disclosure_backend/project/urls.py
@@ -15,9 +15,9 @@ urlpatterns = patterns('',
         'document_root': settings.STATIC_ROOT,
         'show_indexes': True,
     }),
-    url(r'^search/', views.search_view),
     url(r'^locations/(?P<fips_id>[0-9]+)$', views.location_view),
-    url(r'^docs/', include('rest_framework_swagger.urls'))
+    url(r'', include('fips.urls')),
+    url(r'^docs/', include('rest_framework_swagger.urls')),
 )
 
 # Register all API Viewsets:
