+++ fips/serializers.py
@@ -1,0 +1,5 @@
+from rest_framework import serializers
+
+
+class LocationSerializer(serializers.Serializer):
+    pass
